bail_reckoner/ai_service/requirements.txt
flask
python-dotenv
google-cloud-aiplatform
vertexai
gunicorn